requirements.txt
flask
gunicorn
python-dotenv
supabase-py
passlib